# perf: use vector instead CDeterministicMNList for computing rotating quorums