## Issue being fixed or feature implemented
Helper to get quorum members uses CDeterministicMNList as a source of data.
It makes temporary calculation slower, because temporary lists of masternodes should not construct heavy objects CDeterministicMNList just to keep a list of them.


## What was done?
 - helpers `CalculateQuorum`, `CalculateScores` are moved from class `CDeterministicMNList` to `llmq::utils` so far as they don't actually use any private data of `CDeterministicMNList`
 - helper `CalculateQuorum` can accept as source of data not only CDeterministicMNList but std::vector with shared ptrs.
 - helper `BuildNewQuorumQuarterMembers` does not use CDeterministicMNList for `MnsNotUsedAtH`
 - `GetMNUsageBySnapshot` does not use CDeterministicMNList anymore

Also, `BuildNewQuorumQuarterMembers` uses `std::move` when possible.


## How Has This Been Tested?
invalidate + reconsider 15000 blocks; check logs and `perf`.

By perf `PreComputeQuorumMembers` got almost double faster; most improvements came from `GetQuorumQuarterMembersBySnapshot` as expected.
It should give roughly 2% overall improvement for block validation speed and reindex.

PR:
<img width="747" alt="image" src="https://github.com/user-attachments/assets/d27778fa-cc09-4bd7-bccd-601fa6df75f3" />
```
2025-05-20T15:46:53Z [bench]         - m_qblockman: 0.05ms [87.08s]
2025-05-20T15:46:53Z [bench] - Connect block: 21.17ms [249.92s (17.38ms/blk)]
```

Develop:
<img width="747" alt="image" src="https://github.com/user-attachments/assets/aff54059-703d-4c76-967a-15279b1b420c" />
```
2025-05-20T16:11:33Z [bench]         - m_qblockman: 0.05ms [93.42s]
2025-05-20T16:11:33Z [bench] - Connect block: 19.06ms [256.95s (17.13ms/blk)]
```

## Breaking Changes
N/A


## Checklist:
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added or updated relevant unit/integration/functional/e2e tests
- [ ] I have made corresponding changes to the documentation
- [x] I have assigned this pull request to a milestone